Separate CommandDialog from CommandMenu.
