+++ packages/ui/react-ui-editor/src/extensions/command-dialog/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export * from './command-dialog';
+export { type Action } from './action';
