+++ packages/ui/react-ui-editor/src/extensions/command-menu/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './command-menu';
+export * from './useCommandMenu';
