--- packages/ui/react-ui-editor/src/extensions/command/index.ts
@@ -1,10 +1,0 @@
-//
-// Copyright 2024 DXOS.org
-//
-
-export * from './action';
-export * from './command';
-export * from './command-menu';
-export * from './floating-menu';
-export * from './typeahead';
-export * from './useCommandMenu';
