+++ packages/ui/react-ui-editor/src/extensions/index.ts
@@ -8,11 +8,13 @@ export * from './autoscroll';
 export * from './automerge';
 export * from './awareness';
 export * from './blast';
-export * from './command';
+export * from './command-dialog';
+export * from './command-menu';
 export * from './comments';
 export * from './debug';
 export * from './dnd';
 export * from './factories';
+export * from './floating-menu';
 export * from './focus';
 export * from './folding';
 export * from './hashtag';
