+++ packages/ui/react-ui-editor/src/stories/CommandDialog.stories.tsx
@@ -12,7 +12,7 @@ import { mx } from '@dxos/react-ui-theme';
 import { withLayout, withTheme } from '@dxos/storybook-utils';
 
 import { editorWidth } from '../defaults';
-import { type Action, command, floatingMenu } from '../extensions';
+import { type Action, commandDialog } from '../extensions';
 import { str } from '../testing';
 import { createRenderer } from '../util';
 
@@ -66,14 +66,13 @@ const CommandDialog = ({ onAction }: { onAction: (action?: Action) => void }) =>
 };
 
 const meta = {
-  title: 'ui/react-ui-editor/Command',
+  title: 'ui/react-ui-editor/CommandDialog',
   decorators: [withTheme, withLayout({ fullscreen: true })],
   render: () => (
     <EditorStory
       text={str('# Command', '', '', '')}
       extensions={[
-        floatingMenu(),
-        command({
+        commandDialog({
           renderDialog: createRenderer(CommandDialog),
           onHint: () => "Press '/' for commands",
         }),
